account for outs and metrics
